01_LiverGuard_CDSS_System/Dev_Archive/Other_Backups/Django-projec-Final/CDSS/ddi_web/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
